+++ docs/changes.rst
@@ -12,6 +12,7 @@ Not yet released.
 * :ref:`check-reused` is no longer triggered for languages with a single plural form.
 * WebP is now supported for :ref:`screenshots`.
 * Avoid duplicate notification when a user is subscribed to overlapping scopes.
+* Improved OCR support for non-English languages in :ref:`screenshots`.
 
 **Bug fixes**